+++ packages/analytics/dashboard-renderer/src/components/DashboardRenderer.cy.ts
@@ -1,30 +1,70 @@
-// Cypress component test spec file
-
-import DashboardRenderer from './DashboardRenderer.vue'
 import { ChartTypes } from '../types'
 import { ChartMetricDisplay } from '@kong-ui-public/analytics-chart'
-import type { GlobalMountOptions } from '@vue/test-utils/dist/types'
-import { nonTsExploreResponse } from '../../sandbox/mock-data'
-
 import { INJECT_QUERY_PROVIDER } from '../constants'
-import type { AnalyticsBridge } from '@kong-ui-public/analytics-utilities'
-
-const mockQueryProvider: AnalyticsBridge = {
-  queryFn: async () => { return nonTsExploreResponse },
-}
-
-const global: GlobalMountOptions = {
-  provide: {
-    [INJECT_QUERY_PROVIDER]: mockQueryProvider,
-  },
-}
+import type { AnalyticsBridge, ExploreQuery, ExploreResultV4, Timeframe } from '@kong-ui-public/analytics-utilities'
+import {
+  datePickerSelectionToTimeframe,
+  generateSingleMetricTimeSeriesData,
+  generateCrossSectionalData,
+  TimeframeKeys,
+  TimePeriods,
+} from '@kong-ui-public/analytics-utilities'
+import DashboardRenderer from './DashboardRenderer.vue'
+import { nonTsExploreResponse, timeSeriesExploreResponse, summaryDashboardConfig } from '../../sandbox/mock-data'
 
 describe('<DashboardRenderer />', () => {
+  beforeEach(() => {
+    cy.viewport(1200, 1000)
+  })
+
+  const mockQueryProvider = (): AnalyticsBridge => {
+    const queryFn = (query: ExploreQuery): Promise<ExploreResultV4> => {
+      // Dimensions to use if query is not provided
+      const dimensionMap = { statusCode: ['1XX', '2XX', '3XX', '4XX', '5XX'] }
+
+      console.log(' query filters >>> ', query.filters)
+      if (query.dimensions && query.dimensions.findIndex(d => d === 'time') > -1) {
+        if (query.metrics && query.metrics[0] === 'request_count') {
+        // Traffic + Error rate cards
+          return Promise.resolve(nonTsExploreResponse)
+        } else if (query.metrics && query.metrics[0] === 'response_latency_p99' && query.metrics.length === 1) {
+        // Latency metrics card
+          return Promise.resolve(timeSeriesExploreResponse)
+        } else {
+        // Timeseries Line chart
+          const timeSeriesResponse = generateSingleMetricTimeSeriesData(
+            { name: 'TotalRequests', unit: 'count' },
+            { statusCode: query.metrics as string[] },
+          ) as ExploreResultV4
+
+          return Promise.resolve(timeSeriesResponse)
+        }
+      } else {
+      // Bar charts (non-time series)
+        const nonTimeSeriesResponse = generateCrossSectionalData(
+          [
+            { name: 'TotalRequests', unit: 'count' },
+          ],
+          dimensionMap,
+        ) as ExploreResultV4
+
+        return Promise.resolve(nonTimeSeriesResponse)
+      }
+    }
+
+    return {
+      queryFn: cy.spy(queryFn).as('fetcher'),
+    }
+  }
+
   it('Renders the correct number of tiles', () => {
     const props = {
       context: {
         filters: [],
-        timeSpec: '',
+        timeSpec: {
+          type: 'relative',
+          time_range: '15m',
+        },
       },
       config: {
         gridSize: {
@@ -35,18 +75,18 @@ describe('<DashboardRenderer />', () => {
           {
             definition: {
               chart: {
-                type: ChartTypes.HorizontalBar,
+                type: ChartTypes.GoldenSignals,
               },
               query: {},
             },
             layout: {
-              size: {
-                cols: 2,
-                rows: 2,
-              },
               position: {
-                col: 1,
-                row: 1,
+                col: 0,
+                row: 0,
+              },
+              size: {
+                cols: 6,
+                rows: 1,
               },
             },
           },
@@ -58,7 +98,7 @@ describe('<DashboardRenderer />', () => {
                 reverseDataset: true,
                 numerator: 0,
               },
-              query: {},
+              query: { },
             },
             layout: {
               size: {
@@ -77,10 +117,112 @@ describe('<DashboardRenderer />', () => {
 
     cy.mount(DashboardRenderer, {
       props,
-      global,
+      global: {
+        provide: {
+          [INJECT_QUERY_PROVIDER]: mockQueryProvider(),
+        },
+      },
     })
 
     cy.get('.kong-ui-public-dashboard-renderer').should('be.visible')
     cy.get('.tile-boundary').should('have.length', 2)
+
+    // Two queries for the metric cards, one for the gauge chart
+    cy.get('@fetcher').should('have.been.calledThrice')
+  })
+
+  it('Changing the timeframe changes the query', () => {
+    const oneDayTimeframe: Timeframe = TimePeriods.get(TimeframeKeys.ONE_DAY)!
+    const props = {
+      context: {
+        filters: [],
+        timeSpec: oneDayTimeframe.v4Query(),
+      },
+      config: summaryDashboardConfig,
+    }
+
+    cy.mount(DashboardRenderer, {
+      props,
+      global: {
+        provide: {
+          [INJECT_QUERY_PROVIDER]: mockQueryProvider(),
+        },
+      },
+    }).then(({ wrapper }) => {
+      // Two queries for the metric cards, three for the charts
+      cy.get('@fetcher').should('have.callCount', 5)
+
+      cy.get('.kong-ui-public-dashboard-renderer').should('be.visible')
+      cy.get('.tile-boundary').should('have.length', 4)
+
+      cy.get('.metricscard-trend-range').eq(0).should('contain.text', 'previous 24 hours')
+
+      cy.get('@fetcher').should('always.have.been.calledWithMatch', Cypress.sinon.match.hasNested('time_range'))
+      cy.get('@fetcher').should('always.have.been.calledWithMatch', Cypress.sinon.match({
+        time_range: { time_range: '24h' },
+      })).then(() => {
+        cy.get('@fetcher').then((m) => m.resetHistory()).then(() => {
+
+          const sevenDayTimeframe: Timeframe = TimePeriods.get(TimeframeKeys.SEVEN_DAY)!
+
+          wrapper.setProps({
+            context: {
+              filters: [],
+              timeSpec: sevenDayTimeframe.v4Query(),
+            },
+          }).then(() => {
+            // Two more queries for the metric cards, three for the charts
+            cy.get('@fetcher').should('have.callCount', 5)
+
+            cy.get('.kong-ui-public-dashboard-renderer').should('be.visible')
+            cy.get('.tile-boundary').should('have.length', 4)
+
+            cy.get('.metricscard-trend-range').eq(0).should('contain.text', 'previous 7 days')
+
+            cy.get('@fetcher').should('always.have.been.calledWithMatch', Cypress.sinon.match.hasNested('time_range'))
+            cy.get('@fetcher').should('always.have.been.calledWithMatch', Cypress.sinon.match({
+              time_range: { time_range: '7d' },
+            }))
+          })
+        })
+      })
+    })
+  })
+
+  it('Renders a dashboard with custom timeframe, checks query filter', () => {
+    const customTimeframe = datePickerSelectionToTimeframe({
+      timePeriodsKey: 'custom',
+      start: new Date('2024-03-03T21:10:28.969Z'),
+      end: new Date('2024-03-06T21:10:28.969Z'),
+    }) as Timeframe
+
+    const props = {
+      context: {
+        filters: [],
+        timeSpec: customTimeframe.v4Query(),
+      },
+      config: summaryDashboardConfig,
+    }
+
+    cy.mount(DashboardRenderer, {
+      props,
+      global: {
+        provide: {
+          [INJECT_QUERY_PROVIDER]: mockQueryProvider(),
+        },
+      },
+    })
+
+    cy.get('@fetcher').should('always.have.been.calledWithMatch', Cypress.sinon.match({
+      time_range: { type: 'absolute' },
+    }))
+
+    cy.get('@fetcher').should('have.been.calledWithMatch', Cypress.sinon.match({
+      filters: [{
+        dimension: 'control_plane',
+        type: 'in',
+        values: ['default_uuid'],
+      }],
+    }))
   })
 })