+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.11](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@4.0.10...@kong-ui-public/analytics-metric-provider@4.0.11) (2024-03-07)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [4.0.10](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@4.0.9...@kong-ui-public/analytics-metric-provider@4.0.10) (2024-03-07)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
 ## [4.0.9](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@4.0.8...@kong-ui-public/analytics-metric-provider@4.0.9) (2024-03-06)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-metric-provider