+++ packages/core/error-boundary/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.28](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/error-boundary@2.0.27...@kong-ui-public/error-boundary@2.0.28) (2024-03-07)
+
+**Note:** Version bump only for package @kong-ui-public/error-boundary
+
+
+
+
+
 ## [2.0.27](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/error-boundary@2.0.26...@kong-ui-public/error-boundary@2.0.27) (2024-02-26)
 
 **Note:** Version bump only for package @kong-ui-public/error-boundary